# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Filmes",
        "description": "API desenvolvida para uma plataforma de filmes",
        "contact": {
            "email": "josebrunocampanholi@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de testes"
        }
    ],
    "paths": {
...
        "/api/filmes": {
            "get": {
                "tags": [
                    "filmes"
                ],
                "summary": "Filmes cadastrados no sistema",
                "description": "Retorna os filmes cadastrados no sistema paginados em 15",
                "operationId": "indexFilme",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "filmes": {
                                            "$ref": "#/components/schemas/Filme"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "filmes"
                ],
                "summary": "Cria um Filme",
                "description": "Método para criar um novo Filme no sistema",
                "operationId": "storeFilme",
                "requestBody": {
                    "description": "Obrigatório nome do Filme",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Vingadores"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Novo Filme cadastrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro nos parâmetros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Informe todos os dados obrigatórios"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/filmes/{id}": {
            "get": {
                "tags": [
                    "filmes"
                ],
                "summary": "Mostra Filme",
                "description": "Retorna Filme pelo ID da url",
                "operationId": "showFilme",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Filme",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "filme": {
                                            "$ref": "#/components/schemas/Filme"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "filmes"
                ],
                "summary": "Atualiza Filme",
                "description": "Atualiza um Filme, de acordo com o ID",
                "operationId": "updateFilme",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Filme",
                        "required": true,
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Nome do Filme",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Vingadores"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Filme atualizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro nos parâmetros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Informe todos os dados obrigatórios"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "filmes"
                ],
                "summary": "Deleta Filme",
                "description": "Deleta Filme pelo ID da url",
                "operationId": "destroyFilme",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Filme",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Sem conteúdo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/filmes/{id}/categoria/{id_categoria}": {
            "post": {
                "tags": [
                    "filmes"
                ],
                "summary": "Categorizar um filme",
                "description": "Colocar uma categoria em um filme",
                "operationId": "categorizar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Filme",
                        "required": true,
                        "example": "1"
                    },
                    {
                        "name": "id_categoria",
                        "in": "path",
                        "description": "ID da Categoria",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Filme Categorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Parâmetros não encontrados"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/filmes/{id}/ator/{id_ator}": {
            "post": {
                "tags": [
                    "filmes"
                ],
                "summary": "Coloca um ator em um filme",
                "description": "Coloca um ator em um filme",
                "operationId": "colocarAtor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Filme",
                        "required": true,
                        "example": "1"
                    },
                    {
                        "name": "id_ator",
                        "in": "path",
                        "description": "ID do Ator",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Ator inserido no filme"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Parâmetros não encontrados"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/api/filmes/{id}/atores": {
            "get": {
                "tags": [
                    "filmes"
                ],
                "summary": "Atores do Filme",
                "description": "Retorna os atores que o filme tem",
                "operationId": "atoresFilme",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do filme",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "atores": {
                                            "$ref": "#/components/schemas/Ator"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Filme": {
                "required": [
                    "nome"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "nome": {
                        "description": "Nome do Filme",
                        "type": "string",
                        "format": "string",
                        "example": "Vingadores"
                    },
                    "created_at": {
                        "description": "Criação inicial timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Última mudança timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Coloque um Token no formato (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}